+++ test/Shardis.Migration.Tests/SegmentedEnumerationMigrationPlannerTests.cs
@@ -1,5 +1,3 @@
-using AwesomeAssertions;
-
 using Shardis.Migration.Model;
 using Shardis.Migration.Planning;
 using Shardis.Model;