+++ test/Shardis.Migration.Tests/TopologyValidatorTests.cs
@@ -1,5 +1,3 @@
-using AwesomeAssertions;
-
 using Shardis.Migration.Topology;
 using Shardis.Model;
 using Shardis.Persistence;